Sync licnese before checking for updates in Helm mode
