+++ pkg/replicatedapp/api.go
@@ -1,0 +1,74 @@
+package replicatedapp
+
+import (
+	"fmt"
+	"io/ioutil"
+	"net/http"
+
+	"github.com/pkg/errors"
+	kotsv1beta1 "github.com/replicatedhq/kots/kotskinds/apis/kots/v1beta1"
+	"github.com/replicatedhq/kots/kotskinds/client/kotsclientset/scheme"
+	"github.com/replicatedhq/kots/pkg/util"
+)
+
+type LicenseData struct {
+	LicenseBytes []byte
+	License      *kotsv1beta1.License
+}
+
+func GetLatestLicense(license *kotsv1beta1.License) (*LicenseData, error) {
+	url := fmt.Sprintf("%s/license/%s", license.Spec.Endpoint, license.Spec.AppSlug)
+
+	licenseData, err := getLicenseFromAPI(url, license.Spec.LicenseID)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to get license from api")
+	}
+
+	return licenseData, nil
+}
+
+func GetLatestLicenseForHelm(licenseID string) (*LicenseData, error) {
+	url := fmt.Sprintf("%s/license", util.GetReplicatedAPIEndpoint())
+	licenseData, err := getLicenseFromAPI(url, licenseID)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to get helm license from api")
+	}
+
+	return licenseData, nil
+}
+
+func getLicenseFromAPI(url string, licenseID string) (*LicenseData, error) {
+	req, err := util.NewRequest("GET", url, nil)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to call newrequest")
+	}
+
+	req.SetBasicAuth(licenseID, licenseID)
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to execute get request")
+	}
+	defer resp.Body.Close()
+
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to load response")
+	}
+
+	if resp.StatusCode >= 400 {
+		return nil, errors.Errorf("unexpected result from get request: %d, data: %s", resp.StatusCode, body)
+	}
+
+	decode := scheme.Codecs.UniversalDeserializer().Decode
+	obj, _, err := decode(body, nil, nil)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to decode latest license data")
+	}
+
+	data := &LicenseData{
+		LicenseBytes: body,
+		License:      obj.(*kotsv1beta1.License),
+	}
+	return data, nil
+}
